+++ example/lib/main.dart
@@ -1,9 +1,95 @@
+//@dart=2.9
 import 'dart:async';
-
-import 'package:flutter/material.dart';
+import 'dart:io';
 
 import 'package:callkeep/callkeep.dart';
+import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:flutter/material.dart';
 import 'package:uuid/uuid.dart';
+
+/// For fcm background message handler.
+final FlutterCallkeep _callKeep = FlutterCallkeep();
+bool _callKeepInited = false;
+
+/*
+{
+    "uuid": "xxxxx-xxxxx-xxxxx-xxxxx",
+    "caller_id": "+8618612345678",
+    "caller_name": "hello",
+    "caller_id_type": "number", 
+    "has_video": false,
+
+    "extra": {
+        "foo": "bar",
+        "key": "value",
+    }
+}
+*/
+
+Future<dynamic> myBackgroundMessageHandler(Map<String, dynamic> message) {
+  print('backgroundMessage: message => ${message.toString()}');
+  var payload = message['data'];
+  var callerId = payload['caller_id'] as String;
+  var callerNmae = payload['caller_name'] as String;
+  var uuid = payload['uuid'] as String;
+  var hasVideo = payload['has_video'] == "true";
+
+  final callUUID = uuid ?? Uuid().v4();
+  _callKeep.on(CallKeepPerformAnswerCallAction(),
+      (CallKeepPerformAnswerCallAction event) {
+    print(
+        'backgroundMessage: CallKeepPerformAnswerCallAction ${event.callUUID}');
+    _callKeep.startCall(event.callUUID, callerId, callerNmae);
+
+    Timer(const Duration(seconds: 1), () {
+      print(
+          '[setCurrentCallActive] $callUUID, callerId: $callerId, callerName: $callerNmae');
+      _callKeep.setCurrentCallActive(callUUID);
+    });
+    //_callKeep.endCall(event.callUUID);
+  });
+
+  _callKeep.on(CallKeepPerformEndCallAction(),
+      (CallKeepPerformEndCallAction event) {
+    print('backgroundMessage: CallKeepPerformEndCallAction ${event.callUUID}');
+  });
+  if (!_callKeepInited) {
+    _callKeep.setup(<String, dynamic>{
+      'ios': {
+        'appName': 'CallKeepDemo',
+      },
+      'android': {
+        'alertTitle': 'Permissions required',
+        'alertDescription':
+            'This application needs to access your phone accounts',
+        'cancelButton': 'Cancel',
+        'okButton': 'ok',
+      },
+    });
+    _callKeepInited = true;
+  }
+
+  print('backgroundMessage: displayIncomingCall ($callerId)');
+  _callKeep.displayIncomingCall(callUUID, callerId,
+      localizedCallerName: callerNmae, hasVideo: hasVideo);
+  _callKeep.backToForeground();
+  /*
+
+  if (message.containsKey('data')) {
+    // Handle data message
+    final dynamic data = message['data'];
+  }
+
+  if (message.containsKey('notification')) {
+    // Handle notification message
+    final dynamic notification = message['notification'];
+    print('notification => ${notification.toString()}');
+  }
+
+  // Or do other work.
+  */
+  return null;
+}
 
 void main() {
   runApp(MyApp());
@@ -35,8 +121,17 @@ class Call {
 class _MyAppState extends State<HomePage> {
   final FlutterCallkeep _callKeep = FlutterCallkeep();
   Map<String, Call> calls = {};
-
   String newUUID() => Uuid().v4();
+  final FirebaseMessaging _firebaseMessaging = FirebaseMessaging();
+
+  void iOS_Permission() {
+    _firebaseMessaging.requestNotificationPermissions(
+        IosNotificationSettings(sound: true, badge: true, alert: true));
+    _firebaseMessaging.onIosSettingsRegistered
+        .listen((IosNotificationSettings settings) {
+      print('Settings registered: $settings');
+    });
+  }
 
   void removeCall(String callUUID) {
     setState(() {
@@ -176,9 +271,23 @@ class _MyAppState extends State<HomePage> {
         handleType: 'number', hasVideo: false);
   }
 
+  void didDisplayIncomingCall(CallKeepDidDisplayIncomingCall event) {
+    var callUUID = event.callUUID;
+    var number = event.handle;
+    print('[displayIncomingCall] $callUUID number: $number');
+    setState(() {
+      calls[callUUID] = Call(number);
+    });
+  }
+
+  // void onPushKitToken(CallKeepPushKitToken event) {
+  //   print('[onPushKitToken] token => ${event.token}');
+  // }
+
   @override
   void initState() {
     super.initState();
+    _callKeep.on(CallKeepDidDisplayIncomingCall(), didDisplayIncomingCall);
     _callKeep.on(CallKeepPerformAnswerCallAction(), answerCall);
     _callKeep.on(CallKeepDidPerformDTMFAction(), didPerformDTMFAction);
     _callKeep.on(
@@ -187,6 +296,7 @@ class _MyAppState extends State<HomePage> {
     _callKeep.on(
         CallKeepDidPerformSetMutedCallAction(), didPerformSetMutedCallAction);
     _callKeep.on(CallKeepPerformEndCallAction(), endCall);
+    // _callKeep.on(CallKeepPushKitToken(), onPushKitToken);
 
     _callKeep.setup(<String, dynamic>{
       'ios': {
@@ -200,6 +310,42 @@ class _MyAppState extends State<HomePage> {
         'okButton': 'ok',
       },
     });
+
+    if (Platform.isAndroid) {
+      //if (isIOS) iOS_Permission();
+      //  _firebaseMessaging.requestNotificationPermissions();
+
+      _firebaseMessaging.getToken().then((token) {
+        print('[FCM] token => ' + token);
+      });
+
+      _firebaseMessaging.configure(
+        onMessage: (Map<String, dynamic> message) async {
+          print('onMessage: $message');
+          if (message.containsKey('data')) {
+            // Handle data message
+            var payload = message['data'];
+            var callerId = payload['caller_id'] as String;
+            var callerName = payload['caller_name'] as String;
+            var uuid = payload['uuid'] as String;
+            var hasVideo = payload['has_video'] == "true";
+            final callUUID = uuid ?? Uuid().v4();
+            setState(() {
+              calls[callUUID] = Call(callerId);
+            });
+            _callKeep.displayIncomingCall(callUUID, callerId,
+                localizedCallerName: callerName, hasVideo: hasVideo);
+          }
+        },
+        onBackgroundMessage: myBackgroundMessageHandler,
+        onLaunch: (Map<String, dynamic> message) async {
+          print('onLaunch: $message');
+        },
+        onResume: (Map<String, dynamic> message) async {
+          print('onResume: $message');
+        },
+      );
+    }
   }
 
   Widget buildCallingWidgets() {